pyproject.toml
[project]
name = "vita-tg-bot"
version = "0.1.0"
description = "Telegram bot for meal planning with Google Sheets"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.0.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth>=2.25.0",
    "google-api-python-client>=2.100.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
]

[tool.uv]
managed = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
